Made specifying a template case insensitivve
Made a template service that doesn't depend on github. In the future I
might want to allow adding custom templates, in which case the template
service might search multiple sources, and not just GitHub.

I was padding the template string from within the GitHub service for
display, which was idiotic. I moved the padding under
TemplatesCommand.List.